refactor: remove classic table theme
